wallet: extract EnsureKeyScope shared workflow
